Move inventory_item source into src/, include/ directories
