+++ Makefile
@@ -30,11 +30,14 @@ all: bin/game$(EXE_EXTENSION)
 
 
 
-bin/game$(EXE_EXTENSION): obj/game.o obj/logging.o
-	g++ obj/logging.o obj/game.o -o bin/game $(OPENGL_LIB) -l$(ALLEGRO_FLARE_LIB) $(ALLEGRO_LIBS) -L$(ALLEGRO_FLARE_DIR)/lib -L$(ALLEGRO_DIR)/build/lib
+bin/game$(EXE_EXTENSION): obj/inventory_item.o obj/game.o obj/logging.o
+	g++ obj/inventory_item.o obj/logging.o obj/game.o -o bin/game $(OPENGL_LIB) -l$(ALLEGRO_FLARE_LIB) $(ALLEGRO_LIBS) -L$(ALLEGRO_FLARE_DIR)/lib -L$(ALLEGRO_DIR)/build/lib
 
 obj/game.o: main.cpp style_assets.hpp include/logging.hpp inventory.hpp start_screen.hpp script.hpp target_id.hpp navigation.hpp world_render.hpp game/project.hpp game/script_helper.hpp game/load_game_scripts.hpp $(wildcard game/scripts/*)
 	g++ -o $@ -c -std=gnu++11 $< -I$(ALLEGRO_FLARE_DIR)/include -I$(ALLEGRO_DIR)/include -I./include
 
 obj/logging.o: src/logging.cpp include/logging.hpp
+	g++ -o $@ -c -std=gnu++11 $< -I$(ALLEGRO_FLARE_DIR)/include -I$(ALLEGRO_DIR)/include -I./include
+
+obj/inventory_item.o: src/inventory_item.cpp include/logging.hpp
 	g++ -o $@ -c -std=gnu++11 $< -I$(ALLEGRO_FLARE_DIR)/include -I$(ALLEGRO_DIR)/include -I./include
